Use float complex to reduce memory

+++ src/complex_g3.c
@@ -2,21 +2,22 @@
 * \brief Functions for complex_g3.h
 */
 
-#include "decoder.h"
+#include "complex_g3.h"
+#include <math.h>
 
 complex_g3_t complex_from_polar(double r, double theta_radians)
 {
     complex_g3_t result;
 
-    result.re = r * cos(theta_radians);
-    result.im = r * sin(theta_radians);
+    result.re = (float)(r * cos(theta_radians));
+    result.im = (float)(r * sin(theta_radians));
 
     return result;
 }
 
 double complex_magnitude(complex_g3_t c)
 {
-    return sqrt(c.re*c.re + c.im*c.im);
+    return sqrt((double)c.re * c.re + (double)c.im * c.im);
 }
 
 double complex_decibels(complex_g3_t c){
@@ -27,8 +28,8 @@ complex_g3_t complex_add(complex_g3_t left, complex_g3_t right)
 {
     complex_g3_t result;
 
-    result.re = left.re + right.re;
-    result.im = left.im + right.im;
+    result.re = (float)(left.re + right.re);
+    result.im = (float)(left.im + right.im);
 
     return result;
 }
@@ -37,8 +38,8 @@ complex_g3_t complex_sub(complex_g3_t left, complex_g3_t right)
 {
     complex_g3_t result;
 
-    result.re = left.re - right.re;
-    result.im = left.im - right.im;
+    result.re = (float)(left.re - right.re);
+    result.im = (float)(left.im - right.im);
 
     return result;
 }
@@ -47,8 +48,8 @@ complex_g3_t complex_mult(complex_g3_t left, complex_g3_t right)
 {
     complex_g3_t result;
 
-    result.re = left.re*right.re - left.im*right.im;
-    result.im = left.re*right.im + left.im*right.re;
+    result.re = (float)(left.re*right.re - left.im*right.im);
+    result.im = (float)(left.re*right.im + left.im*right.re);
 
     return result;
 }
